[Readme] Write a better example using KeepRunningBatch [Headers] Mark 0-as-nullptr with comment, rename to BenchmarkTimeDescription as per review [Warnings] Remove downcast warning from benchmark runner

+++ src/benchmark_runner.cc
@@ -254,8 +254,8 @@ class BenchmarkRunner {
     if (multiplier <= 1.0) multiplier = 2.0;
 
     // So what seems to be the sufficiently-large iteration count? Round up.
-    const size_t max_next_iters =
-        0.5 + std::max(multiplier * i.iters, i.iters + 1.0);
+    const size_t max_next_iters = static_cast<size_t>(
+        0.5 + std::max(multiplier * i.iters, i.iters + 1.0));
     // But we do have *some* sanity limits though..
     const size_t next_iters = std::min(max_next_iters, kMaxIterations);
